Abstracted several functions to class called DynamicArray, working but more work to do, got to move shift functions from arraylist to dynamic array
