+++ .idea/modules/java_data_structures_test.iml
@@ -34,6 +34,7 @@
     <orderEntry type="library" name="Gradle: commons-logging:commons-logging:1.2" level="project" />
     <orderEntry type="library" name="Gradle: commons-collections:commons-collections:3.2.2" level="project" />
     <orderEntry type="library" name="Gradle: aopalliance:aopalliance:1.0" level="project" />
+    <orderEntry type="library" name="Python 2.7 interpreter library" level="application" />
   </component>
   <component name="TestModuleProperties" production-module="java_data_structures_main" />
 </module>
