+++ data/csv/avg_cost_to_insert_into_arraylist.csv
@@ -1,300 +1,300 @@
 2,1
 3,2
-4,1
+4,3
 5,4
-6,4
+6,2
 7,1
-8,6
-9,6
+8,1
+9,7
 10,8
 11,0
-12,4
-13,6
-14,11
-15,10
-16,9
-17,12
-18,12
-19,18
+12,3
+13,3
+14,3
+15,4
+16,6
+17,7
+18,14
+19,11
 20,0
-21,15
-22,11
-23,19
-24,9
-25,12
-26,23
-27,22
-28,1
-29,28
-30,3
-31,26
-32,27
-33,24
-34,11
-35,7
+21,9
+22,9
+23,8
+24,20
+25,3
+26,4
+27,12
+28,5
+29,11
+30,29
+31,22
+32,24
+33,5
+34,10
+35,32
 36,0
 37,31
-38,18
-39,34
-40,20
-41,37
-42,7
-43,18
-44,2
-45,18
-46,16
-47,35
-48,17
-49,15
-50,27
-51,5
-52,4
-53,37
-54,15
-55,51
-56,33
-57,28
-58,36
-59,20
-60,14
-61,48
-62,6
-63,26
-64,9
+38,21
+39,28
+40,26
+41,6
+42,14
+43,31
+44,37
+45,3
+46,2
+47,19
+48,40
+49,40
+50,43
+51,3
+52,44
+53,42
+54,46
+55,27
+56,29
+57,43
+58,46
+59,52
+60,10
+61,32
+62,54
+63,13
+64,11
 65,0
-66,50
-67,61
-68,17
-69,47
-70,8
-71,48
-72,59
-73,56
-74,35
-75,18
-76,39
-77,68
-78,23
-79,8
-80,59
-81,59
-82,13
-83,64
-84,44
-85,49
-86,14
-87,11
-88,30
-89,68
-90,13
-91,56
-92,83
-93,57
-94,54
-95,41
-96,9
-97,54
-98,1
-99,25
-100,74
-101,92
-102,15
-103,23
-104,76
-105,13
-106,27
-107,68
-108,59
-109,54
-110,63
-111,106
-112,55
-113,26
-114,42
-115,90
-116,22
+66,65
+67,4
+68,12
+69,28
+70,61
+71,66
+72,60
+73,36
+74,71
+75,71
+76,8
+77,70
+78,46
+79,76
+80,50
+81,14
+82,61
+83,11
+84,7
+85,76
+86,5
+87,15
+88,48
+89,70
+90,29
+91,63
+92,60
+93,9
+94,35
+95,79
+96,24
+97,80
+98,20
+99,35
+100,53
+101,15
+102,89
+103,98
+104,75
+105,75
+106,53
+107,97
+108,103
+109,92
+110,38
+111,75
+112,29
+113,24
+114,27
+115,86
+116,115
 117,0
-118,107
-119,41
-120,73
-121,113
-122,72
-123,121
-124,18
-125,74
-126,109
-127,13
-128,50
-129,35
-130,31
-131,61
-132,85
-133,95
-134,84
-135,132
-136,9
-137,35
-138,77
-139,15
-140,29
-141,131
-142,58
-143,140
-144,86
-145,5
-146,58
-147,116
-148,6
-149,92
-150,40
-151,118
-152,151
-153,106
-154,36
-155,136
-156,141
-157,63
-158,52
-159,73
-160,105
-161,147
-162,119
-163,56
-164,62
-165,32
-166,25
-167,108
-168,82
-169,124
-170,156
-171,7
-172,27
-173,12
-174,92
-175,117
-176,109
-177,114
-178,131
-179,147
-180,142
-181,66
-182,7
-183,123
-184,66
-185,166
-186,179
-187,102
-188,179
-189,70
-190,106
-191,101
-192,100
-193,66
-194,40
-195,131
-196,20
-197,60
-198,101
-199,7
-200,6
-201,91
-202,110
-203,3
-204,141
-205,109
-206,70
-207,127
-208,56
-209,188
+118,79
+119,63
+120,38
+121,111
+122,68
+123,36
+124,56
+125,106
+126,106
+127,3
+128,73
+129,102
+130,50
+131,130
+132,112
+133,80
+134,117
+135,4
+136,76
+137,30
+138,20
+139,32
+140,110
+141,3
+142,95
+143,10
+144,56
+145,103
+146,57
+147,79
+148,138
+149,38
+150,21
+151,104
+152,5
+153,149
+154,71
+155,24
+156,48
+157,111
+158,47
+159,15
+160,65
+161,37
+162,132
+163,81
+164,96
+165,8
+166,26
+167,125
+168,106
+169,136
+170,128
+171,45
+172,160
+173,13
+174,70
+175,4
+176,38
+177,5
+178,47
+179,110
+180,106
+181,69
+182,144
+183,36
+184,32
+185,112
+186,99
+187,137
+188,156
+189,31
+190,140
+191,182
+192,72
+193,148
+194,13
+195,160
+196,75
+197,69
+198,17
+199,152
+200,58
+201,23
+202,31
+203,122
+204,24
+205,93
+206,112
+207,43
+208,100
+209,123
 210,0
-211,204
-212,163
-213,69
-214,68
-215,175
-216,209
-217,78
-218,158
-219,89
-220,173
-221,214
-222,166
-223,219
-224,23
-225,142
-226,52
-227,77
-228,221
-229,107
-230,68
-231,168
-232,155
-233,82
-234,3
-235,203
-236,133
-237,53
-238,206
-239,87
-240,236
-241,110
-242,191
-243,142
-244,124
-245,234
-246,64
-247,77
-248,14
-249,191
-250,138
-251,49
-252,30
-253,167
-254,250
-255,171
-256,90
-257,127
-258,166
-259,48
-260,153
-261,43
-262,249
-263,34
-264,170
-265,53
-266,247
-267,98
-268,233
-269,25
-270,214
-271,134
-272,248
-273,125
-274,32
-275,111
-276,35
-277,259
-278,145
-279,90
-280,161
-281,32
-282,69
-283,155
-284,29
-285,259
-286,63
-287,195
-288,43
-289,92
-290,207
-291,73
-292,45
-293,210
-294,7
-295,42
-296,83
-297,169
-298,69
-299,129
-300,153
-301,149
+211,17
+212,151
+213,184
+214,88
+215,58
+216,163
+217,113
+218,50
+219,50
+220,22
+221,158
+222,129
+223,87
+224,48
+225,208
+226,105
+227,26
+228,164
+229,158
+230,82
+231,103
+232,206
+233,155
+234,61
+235,13
+236,31
+237,157
+238,149
+239,55
+240,190
+241,125
+242,57
+243,41
+244,103
+245,82
+246,135
+247,244
+248,24
+249,25
+250,94
+251,33
+252,87
+253,62
+254,5
+255,92
+256,8
+257,95
+258,39
+259,21
+260,237
+261,83
+262,41
+263,259
+264,175
+265,166
+266,257
+267,156
+268,187
+269,184
+270,169
+271,29
+272,235
+273,94
+274,11
+275,240
+276,70
+277,224
+278,120
+279,210
+280,222
+281,168
+282,208
+283,2
+284,265
+285,200
+286,285
+287,135
+288,29
+289,58
+290,206
+291,27
+292,36
+293,259
+294,91
+295,80
+296,45
+297,294
+298,96
+299,267
+300,209
+301,175
