+++ data/csv/avg_cost_to_remove_from_arraylist.csv
@@ -1,300 +1,300 @@
-300,38
-299,87
-298,116
-297,138
-296,167
-295,106
-294,261
-293,250
-292,232
-291,274
-290,147
-289,108
-288,53
-287,212
-286,155
-285,256
-284,210
-283,49
-282,231
-281,280
-280,231
-279,265
-278,65
-277,135
-276,263
-275,184
-274,162
-273,250
-272,159
-271,176
-270,61
-269,117
-268,85
-267,41
-266,185
-265,183
-264,224
-263,232
-262,14
-261,227
-260,83
-259,152
-258,250
-257,168
-256,103
-255,39
-254,138
-253,238
-252,220
-251,5
-250,100
-249,74
-248,56
-247,42
-246,139
-245,55
-244,42
-243,19
-242,128
-241,145
-240,228
-239,126
-238,108
-237,98
-236,132
-235,129
-234,134
-233,172
-232,81
-231,149
-230,140
-229,115
-228,125
-227,126
-226,30
-225,185
-224,220
-223,114
-222,169
-221,47
-220,73
-219,186
-218,110
-217,39
-216,192
-215,75
-214,41
-213,53
-212,168
-211,79
-210,208
-209,64
-208,120
-207,1
-206,119
-205,17
-204,62
-203,65
-202,60
-201,10
-200,33
-199,167
-198,121
-197,121
-196,17
-195,183
-194,140
-193,96
-192,187
-191,62
-190,170
-189,48
-188,178
-187,75
-186,10
-185,171
-184,10
-183,8
-182,13
-181,57
-180,2
-179,179
-178,84
-177,105
-176,127
-175,158
-174,24
-173,60
-172,169
-171,80
-170,67
-169,121
-168,113
-167,74
-166,100
-165,68
-164,3
-163,58
-162,68
-161,47
-160,23
-159,133
-158,106
-157,21
-156,45
-155,86
-154,77
-153,53
-152,73
-151,103
-150,113
-149,1
-148,120
-147,137
-146,66
-145,120
-144,140
-143,7
-142,57
-141,60
-140,96
-139,27
-138,17
-137,1
-136,132
-135,71
-134,26
-133,91
-132,125
-131,90
-130,121
-129,69
-128,66
-127,77
-126,66
-125,51
-124,110
-123,97
-122,52
-121,23
-120,89
-119,92
-118,57
-117,87
-116,101
-115,14
+300,232
+299,182
+298,15
+297,223
+296,161
+295,271
+294,184
+293,259
+292,36
+291,267
+290,150
+289,85
+288,112
+287,217
+286,143
+285,194
+284,23
+283,139
+282,127
+281,199
+280,7
+279,165
+278,186
+277,154
+276,246
+275,65
+274,221
+273,273
+272,22
+271,131
+270,218
+269,28
+268,216
+267,77
+266,264
+265,76
+264,194
+263,14
+262,208
+261,259
+260,216
+259,67
+258,148
+257,2
+256,50
+255,6
+254,179
+253,44
+252,135
+251,125
+250,216
+249,32
+248,100
+247,217
+246,229
+245,138
+244,164
+243,90
+242,219
+241,149
+240,151
+239,153
+238,214
+237,181
+236,104
+235,60
+234,99
+233,118
+232,89
+231,146
+230,23
+229,19
+228,137
+227,61
+226,17
+225,43
+224,74
+223,66
+222,135
+221,75
+220,149
+219,117
+218,103
+217,186
+216,97
+215,142
+214,66
+213,8
+212,199
+211,53
+210,206
+209,194
+208,138
+207,62
+206,94
+205,25
+204,174
+203,111
+202,85
+201,84
+200,184
+199,36
+198,3
+197,117
+196,26
+195,29
+194,48
+193,138
+192,136
+191,111
+190,158
+189,50
+188,26
+187,154
+186,158
+185,139
+184,97
+183,152
+182,6
+181,129
+180,150
+179,111
+178,59
+177,141
+176,35
+175,145
+174,160
+173,131
+172,97
+171,105
+170,89
+169,99
+168,4
+167,160
+166,35
+165,159
+164,58
+163,142
+162,145
+161,21
+160,34
+159,63
+158,34
+157,129
+156,122
+155,26
+154,62
+153,25
+152,14
+151,112
+150,45
+149,28
+148,37
+147,18
+146,81
+145,45
+144,31
+143,140
+142,135
+141,93
+140,2
+139,129
+138,81
+137,122
+136,27
+135,19
+134,13
+133,30
+132,12
+131,120
+130,81
+129,104
+128,111
+127,12
+126,119
+125,56
+124,117
+123,51
+122,54
+121,121
+120,118
+119,82
+118,46
+117,52
+116,9
+115,58
 114,94
-113,18
-112,45
-111,76
-110,10
-109,56
-108,51
-107,97
-106,94
-105,38
-104,31
-103,26
-102,82
-101,5
-100,2
-99,8
-98,57
-97,17
-96,28
-95,23
-94,15
-93,60
-92,17
-91,59
-90,80
-89,36
-88,7
-87,67
-86,11
-85,56
-84,53
-83,69
-82,8
-81,65
-80,36
-79,46
-78,54
-77,13
-76,4
-75,46
-74,27
-73,26
-72,46
-71,71
-70,29
-69,28
-68,26
-67,63
-66,47
-65,51
-64,27
-63,28
-62,11
-61,42
-60,21
-59,6
-58,28
-57,14
-56,11
-55,6
-54,52
-53,32
-52,8
-51,26
-50,19
-49,14
+113,24
+112,19
+111,94
+110,101
+109,81
+108,14
+107,92
+106,71
+105,17
+104,79
+103,31
+102,26
+101,7
+100,90
+99,32
+98,78
+97,90
+96,87
+95,84
+94,33
+93,33
+92,52
+91,13
+90,86
+89,8
+88,11
+87,41
+86,84
+85,16
+84,49
+83,35
+82,74
+81,62
+80,32
+79,72
+78,15
+77,49
+76,39
+75,74
+74,65
+73,1
+72,31
+71,6
+70,15
+69,68
+68,35
+67,39
+66,64
+65,10
+64,15
+63,17
+62,33
+61,31
+60,22
+59,27
+58,25
+57,1
+56,35
+55,55
+54,15
+53,3
+52,12
+51,34
+50,46
+49,33
 48,12
-47,29
-46,32
-45,25
-44,5
-43,6
-42,42
-41,2
-40,35
-39,6
-38,13
-37,16
-36,9
-35,33
-34,17
-33,25
-32,1
-31,22
-30,18
-29,14
-28,24
-27,26
-26,6
-25,19
-24,14
-23,9
-22,1
-21,12
-20,3
+47,42
+46,7
+45,32
+44,28
+43,17
+42,34
+41,8
+40,16
+39,28
+38,28
+37,28
+36,7
+35,34
+34,19
+33,19
+32,18
+31,28
+30,28
+29,2
+28,14
+27,15
+26,17
+25,17
+24,7
+23,13
+22,18
+21,21
+20,5
 19,7
-18,18
-17,11
-16,16
-15,7
-14,14
-13,9
-12,2
-11,2
-10,7
-9,6
-8,7
-7,3
-6,2
-5,4
-4,2
+18,12
+17,10
+16,3
+15,11
+14,8
+13,10
+12,6
+11,10
+10,3
+9,1
+8,5
+7,6
+6,5
+5,5
+4,1
 3,3
 2,1
 1,1
