+++ data/csv/levels_in_avl_tree_vs_n_random_keys.csv
@@ -9,17 +9,17 @@
 9,4
 10,4
 11,4
-12,5
-13,5
+12,4
+13,4
 14,5
 15,5
 16,5
 17,5
 18,5
-19,6
-20,6
-21,6
-22,6
+19,5
+20,5
+21,5
+22,5
 23,6
 24,6
 25,6
@@ -34,14 +34,14 @@
 34,6
 35,6
 36,6
-37,6
-38,6
-39,6
-40,6
-41,6
-42,6
-43,6
-44,6
+37,7
+38,7
+39,7
+40,7
+41,7
+42,7
+43,7
+44,7
 45,7
 46,7
 47,7
@@ -73,7 +73,7 @@
 73,7
 74,7
 75,7
-76,8
+76,7
 77,8
 78,8
 79,8
