+++ data/csv/levels_in_bst_tree_vs_n_random_keys.csv
@@ -2,33 +2,33 @@
 2,2
 3,2
 4,3
-5,3
-6,4
-7,4
+5,4
+6,5
+7,5
 8,5
 9,5
 10,5
-11,6
-12,7
-13,7
-14,8
-15,8
-16,8
+11,5
+12,5
+13,5
+14,6
+15,6
+16,7
 17,8
 18,8
 19,8
 20,8
 21,8
 22,8
-23,8
-24,8
-25,8
-26,8
-27,8
-28,8
-29,8
-30,8
-31,8
+23,9
+24,9
+25,9
+26,9
+27,9
+28,9
+29,9
+30,9
+31,9
 32,9
 33,9
 34,9
@@ -38,9 +38,9 @@
 38,9
 39,9
 40,9
-41,9
-42,9
-43,9
+41,10
+42,10
+43,10
 44,10
 45,10
 46,10
@@ -50,19 +50,19 @@
 50,10
 51,10
 52,10
-53,10
-54,10
-55,10
-56,10
-57,10
-58,10
-59,10
-60,10
-61,10
-62,10
-63,10
-64,10
-65,10
+53,11
+54,11
+55,11
+56,11
+57,11
+58,11
+59,11
+60,11
+61,11
+62,11
+63,11
+64,11
+65,11
 66,11
 67,11
 68,11
@@ -72,29 +72,29 @@
 72,11
 73,11
 74,11
-75,11
-76,11
-77,11
-78,11
-79,11
-80,11
-81,11
-82,11
-83,11
-84,11
-85,11
-86,11
-87,11
-88,11
-89,11
-90,11
-91,11
-92,11
-93,11
-94,11
+75,12
+76,12
+77,12
+78,12
+79,12
+80,12
+81,12
+82,12
+83,12
+84,12
+85,12
+86,12
+87,12
+88,12
+89,12
+90,12
+91,12
+92,12
+93,12
+94,12
 95,12
 96,12
-97,13
-98,13
-99,13
-100,13
+97,12
+98,12
+99,12
+100,12
