+++ data/csv/shifts_required_when_inserting_rand_index.csv
@@ -1,300 +1,300 @@
 2,1
-3,2
-4,5
-5,7
-6,11
-7,15
+3,3
+4,6
+5,8
+6,12
+7,16
 8,21
-9,22
-10,24
-11,24
-12,25
-13,31
-14,35
-15,38
-16,49
-17,58
-18,67
-19,80
-20,80
-21,100
-22,106
-23,112
+9,29
+10,37
+11,37
+12,47
+13,52
+14,65
+15,77
+16,80
+17,93
+18,102
+19,110
+20,110
+21,123
+22,124
+23,125
 24,127
-25,139
-26,160
-27,186
-28,190
-29,217
-30,232
-31,233
-32,251
-33,263
-34,286
-35,312
-36,312
-37,344
-38,377
-39,401
-40,439
-41,448
-42,464
-43,481
-44,491
-45,492
-46,523
-47,532
-48,542
-49,556
-50,568
-51,593
-52,598
-53,614
-54,634
-55,639
-56,668
-57,673
-58,692
-59,747
-60,804
-61,831
-62,845
-63,894
-64,923
-65,923
-66,945
-67,950
-68,1006
-69,1058
-70,1086
-71,1129
-72,1179
-73,1223
-74,1276
-75,1331
-76,1405
-77,1474
-78,1541
-79,1586
-80,1630
-81,1677
-82,1717
-83,1760
-84,1832
-85,1888
-86,1971
-87,2044
-88,2089
-89,2090
-90,2110
-91,2173
-92,2212
-93,2285
-94,2345
-95,2412
-96,2430
-97,2433
-98,2511
-99,2565
-100,2582
-101,2630
-102,2696
-103,2775
-104,2809
-105,2861
-106,2917
-107,2990
-108,3092
-109,3138
-110,3171
-111,3182
-112,3242
-113,3343
-114,3432
-115,3442
-116,3448
-117,3448
-118,3523
-119,3524
-120,3525
-121,3611
-122,3715
-123,3779
-124,3791
-125,3899
-126,3925
-127,3963
-128,3985
-129,4003
-130,4017
-131,4127
-132,4231
-133,4266
-134,4289
-135,4296
-136,4349
-137,4483
-138,4552
-139,4608
-140,4673
-141,4724
-142,4742
-143,4808
-144,4929
-145,4978
-146,5049
-147,5089
-148,5219
-149,5292
-150,5315
-151,5463
-152,5590
-153,5680
-154,5796
-155,5798
-156,5940
-157,5956
-158,6084
-159,6168
-160,6243
-161,6290
-162,6451
-163,6547
-164,6631
-165,6710
-166,6770
-167,6869
-168,6874
-169,6924
-170,7065
-171,7201
-172,7327
-173,7499
-174,7623
-175,7657
-176,7710
-177,7844
-178,7945
-179,7987
-180,8085
-181,8162
-182,8245
-183,8393
-184,8470
-185,8508
-186,8525
-187,8634
-188,8673
-189,8736
-190,8813
-191,8916
-192,9009
-193,9194
-194,9218
-195,9322
-196,9420
-197,9589
-198,9626
-199,9753
-200,9891
-201,10015
-202,10161
-203,10221
-204,10308
-205,10404
-206,10493
-207,10500
-208,10533
-209,10731
-210,10731
-211,10762
-212,10822
-213,10933
-214,11090
-215,11141
-216,11302
-217,11319
-218,11449
-219,11609
-220,11704
-221,11786
-222,11852
-223,11969
-224,12058
-225,12265
-226,12449
-227,12466
-228,12685
-229,12843
-230,12895
-231,13012
-232,13085
-233,13110
-234,13179
-235,13367
-236,13543
-237,13768
-238,13831
-239,13935
-240,14026
-241,14162
-242,14175
-243,14236
-244,14373
-245,14550
-246,14700
-247,14926
-248,15134
-249,15380
-250,15422
-251,15435
-252,15605
-253,15723
-254,15822
-255,15873
-256,16115
-257,16265
-258,16501
-259,16648
-260,16736
-261,16882
-262,16926
-263,17174
-264,17424
-265,17576
-266,17625
-267,17824
-268,17904
-269,17954
-270,18211
-271,18397
-272,18652
-273,18857
-274,19042
-275,19192
-276,19404
-277,19438
-278,19478
-279,19535
-280,19770
-281,19977
-282,20168
-283,20440
-284,20692
-285,20841
-286,20946
-287,21006
-288,21050
-289,21118
-290,21401
-291,21497
-292,21678
-293,21934
-294,22216
-295,22223
-296,22503
-297,22736
-298,22748
-299,22823
-300,23081
-301,23201
+25,146
+26,150
+27,152
+28,173
+29,177
+30,193
+31,211
+32,232
+33,257
+34,274
+35,288
+36,288
+37,303
+38,309
+39,344
+40,382
+41,399
+42,430
+43,463
+44,493
+45,537
+46,561
+47,603
+48,623
+49,661
+50,698
+51,703
+52,751
+53,778
+54,821
+55,846
+56,875
+57,880
+58,926
+59,955
+60,1004
+61,1020
+62,1062
+63,1108
+64,1154
+65,1154
+66,1170
+67,1181
+68,1236
+69,1259
+70,1306
+71,1353
+72,1420
+73,1440
+74,1497
+75,1510
+76,1584
+77,1599
+78,1645
+79,1695
+80,1731
+81,1792
+82,1865
+83,1915
+84,1994
+85,2011
+86,2025
+87,2095
+88,2182
+89,2270
+90,2318
+91,2361
+92,2396
+93,2400
+94,2481
+95,2489
+96,2578
+97,2658
+98,2662
+99,2679
+100,2721
+101,2758
+102,2777
+103,2796
+104,2812
+105,2828
+106,2895
+107,2975
+108,3033
+109,3063
+110,3073
+111,3169
+112,3182
+113,3259
+114,3370
+115,3456
+116,3568
+117,3568
+118,3620
+119,3698
+120,3811
+121,3845
+122,3848
+123,3855
+124,3877
+125,3992
+126,4085
+127,4102
+128,4123
+129,4218
+130,4221
+131,4326
+132,4353
+133,4473
+134,4502
+135,4519
+136,4560
+137,4576
+138,4685
+139,4728
+140,4777
+141,4780
+142,4843
+143,4868
+144,5003
+145,5128
+146,5132
+147,5182
+148,5303
+149,5383
+150,5425
+151,5469
+152,5575
+153,5713
+154,5730
+155,5756
+156,5896
+157,5994
+158,6124
+159,6151
+160,6242
+161,6365
+162,6389
+163,6442
+164,6491
+165,6635
+166,6702
+167,6726
+168,6770
+169,6847
+170,6852
+171,6991
+172,7141
+173,7264
+174,7351
+175,7418
+176,7581
+177,7633
+178,7721
+179,7794
+180,7833
+181,7872
+182,8032
+183,8145
+184,8151
+185,8273
+186,8421
+187,8576
+188,8627
+189,8734
+190,8770
+191,8846
+192,8939
+193,8956
+194,9065
+195,9112
+196,9164
+197,9355
+198,9463
+199,9608
+200,9762
+201,9776
+202,9834
+203,10009
+204,10022
+205,10188
+206,10265
+207,10418
+208,10583
+209,10696
+210,10696
+211,10778
+212,10943
+213,11078
+214,11284
+215,11366
+216,11449
+217,11555
+218,11598
+219,11773
+220,11790
+221,11804
+222,11879
+223,12046
+224,12130
+225,12281
+226,12450
+227,12544
+228,12605
+229,12795
+230,12855
+231,12962
+232,12999
+233,13188
+234,13206
+235,13327
+236,13423
+237,13635
+238,13724
+239,13728
+240,13768
+241,13888
+242,13912
+243,13930
+244,14153
+245,14357
+246,14484
+247,14717
+248,14901
+249,15003
+250,15168
+251,15290
+252,15490
+253,15568
+254,15744
+255,15859
+256,16014
+257,16172
+258,16349
+259,16404
+260,16474
+261,16616
+262,16624
+263,16650
+264,16706
+265,16813
+266,16963
+267,17229
+268,17348
+269,17487
+270,17584
+271,17605
+272,17768
+273,17975
+274,18173
+275,18349
+276,18491
+277,18613
+278,18742
+279,18919
+280,19032
+281,19073
+282,19341
+283,19343
+284,19417
+285,19545
+286,19805
+287,19935
+288,20039
+289,20148
+290,20176
+291,20350
+292,20533
+293,20707
+294,20837
+295,20915
+296,21130
+297,21305
+298,21428
+299,21633
+300,21760
+301,21795
