+++ src/main/java/structures/vectors/ArrayList.java
@@ -11,56 +11,8 @@ import java.util.Collection;
  * @author Jabari Dash
  * @param <E> Generic type
  */
-public final class ArrayList<E> implements DynamicArray<E>, List<E> {
+public final class ArrayList<E> extends DynamicArray<E> implements List<E> {
 
-    /**
-     * The threshold for resizing the internal array. If the
-     * internal array gets 90% full, we will allocate more space
-     * before adding elements. This variable sets that threshold
-     * for resizing. We don't want a threshold that's too small
-     * because at any point in execution, if we have k elements,
-     * we will also have at least k free spaces allocated.
-     *
-     */
-    private static final double RESIZE_THRESHOLD = 0.9;
-
-    /**
-     * The default size of the internal array is 10.
-     */
-    private static final int DEFAULT_SIZE = 10;
-
-    /**
-     * Number of elements present in the ArrayList.
-     */
-    private int size;
-
-    /**
-     * The internal array containing the elements in the ArrayList.
-     */
-    private E[] elements;
-
-    /**
-     * The amount of times the array needed to be resized.
-     * We use this value for analytic purposes. It has no
-     * effect on the structures.performance or behavior of the ArrayList.
-     */
-    private int allocations;
-
-    /**
-     * The number of times an element has been shifted in the array.
-     * This goes for individual elements. So a shift left when the ArrayList
-     * of size 10 will increase shifts by 10. This variable is
-     * used for analytics purposes. It has not effect on
-     * the structures.performance or behavior of the ArrayList.
-     */
-    private int shifts;
-
-    /**
-     * The number of times any element has been copied from one
-     * array to another. This occurs when the {@code copy()} function is
-     * called. If n values are copied, this value increases by n.
-     */
-    private int copies;
 
     /**
      * Instantiates an ArrayList with an internal array of
@@ -76,9 +28,7 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
      */
     @SuppressWarnings("unchecked")
     public ArrayList(int length) {
-
-        // Create a new array no smaller than the default size
-        elements = (E[]) new Object[length < DEFAULT_SIZE ? DEFAULT_SIZE : length];
+        super(length);
     }
 
     /**
@@ -86,7 +36,7 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
      * of the ArrayList is 10.
      */
     public ArrayList() {
-        this(DEFAULT_SIZE);
+        super();
     }
 
     /**
@@ -111,69 +61,6 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
     }
 
     /**
-     * Returns the total numbers of values that
-     * have been copied from one array to another
-     * for the ArrayList to be in its current state.
-     *
-     * @return Number of data copies.
-     */
-    @SuppressWarnings("unused")
-    public int copies() {
-        return copies;
-    }
-
-    /**
-     * Copy a specified range of elements from one
-     * array to another.
-     *
-     * @param start First index in range from which to copy
-     * @param stop Last index in range from which to copy
-     * @param src Source array from which to copy
-     * @param dst Destination array to copy to
-     */
-    private void copy(int start, int stop, int offset, E[] src, E[] dst) {
-
-        // Must be within bounds
-        if (start < 0 || start >= src.length) {
-            throw new IllegalArgumentException("Start index must be in bounds or source array");
-        }
-
-        // If start index exceeds stop index
-        if (start > stop) {
-            throw new IllegalArgumentException("Start index must be less than or equal to stop index");
-        }
-
-        // Must be within bounds
-        // TODO - IntelliJ says stop always < 0, maybe if-else can be simplified
-        if (stop < 0 || stop >= src.length) {
-            throw new IllegalArgumentException("Stop index must be in bounds or src array");
-        }
-
-        // destination array must be at least size of source
-        if (src.length >= dst.length) {
-            throw new IllegalArgumentException("Destination array must be at least the length of the source array");
-        }
-
-        // Copy the array
-        for (int i = start; i <= stop; i++){
-
-            dst[i+offset] = src[i];
-            copies++;
-        }
-
-    }
-
-    /**
-     * Returns the numbers of memory re-allocations that
-     * have occurred to create the ArrayList in its present state.
-     *
-     * @return Number of re-allocations.
-     */
-    public int allocations() {
-        return allocations;
-    }
-
-    /**
      * Determines whether or not this ArrayList is equal to
      * a provided object.
      *
@@ -189,24 +76,6 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
         return this == object || (object instanceof ArrayList && equivalentTo(object));
     }
 
-    /**
-     * If the number of elements in the ArrayList exceeds a specified
-     * threshold (90%), we consider the internal array "full".
-     *
-     * @return True if and only if the capacity threshold is passed.
-     */
-    private boolean full() {
-
-        // Convert size and length to doubles
-        // to perform division with decimals
-        double s = (double) size;
-        double l = (double) elements.length;
-
-        // If the ratio of elements in the DataStructure exceeds the
-        // threshold (default 90%), then we need to allocate more space
-        // in the internal array
-        return (s / l) > RESIZE_THRESHOLD;
-    }
 
     /**
      * Returns the value at a specified index.
@@ -297,18 +166,7 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
         return true;
     }
 
-    /**
-     * Returns the size of the internal list. Note, this
-     * does not specify the number of elements in the list,
-     * rather the number of spaces that the dynamically
-     * allocated array that is used to contain the elements
-     * has at a given point in time.
-     *
-     * @return Length of internal array that contains the elements
-     */
-    public int internalSize() {
-        return elements.length;
-    }
+
 
     /**
      * Inserts a value into the front of the list.
@@ -460,47 +318,6 @@ public final class ArrayList<E> implements DynamicArray<E>, List<E> {
         }
     }
 
-    /**
-     * Returns total number of values that
-     * have been shifted to the left or right
-     * such that the ArrayList is in its present state.
-     *
-     * @return Total number of shifts.
-     */
-    @SuppressWarnings("unused")
-    public int shifts() {
-        return shifts;
-    }
-
-    /**
-     * Returns the number of elements in the ArrayList.
-     *
-     * @return Number of elements in the ArrayList.
-     */
-    @Override
-    public int size() {
-        return size;
-    }
-
-    /**
-     * Returns data structure as an array of specified type.
-     *
-     * @param array Array that specifies the type of the array to return.
-     * @return Array representation of the data structure.
-     */
-    public E[] toArray(E[] array) {
-
-        try {
-
-            array = (E[]) Arrays.copyOf(elements, size(), array.getClass());
-
-        } catch (ArrayStoreException exception) {
-
-            throw new IllegalArgumentException("Array type " + array.getClass().getSimpleName() + " is invalid");
-        }
-
-        return array;
-    }
 
     /**
      * Returns a String representation of the list.
